Add "updateable-rec" class to SODT.
This marks adds "Future updates to this Recommendation may incorporate new features.", Where, "new features" is a link to the process document: https://www.w3.org/2023/Process-20231103/#allow-new-features.

For a recommendation to be updatable, it MUST include this reference, The SODT template doesn't really allow changing the wording of this paragraph.