Pruning the prior art references
(As agreed in https://github.com/w3c/rdf-canon/pull/199#issuecomment-1973415321 bringing this as a separate PR.)

As argued in https://github.com/w3c/rdf-canon/pull/199#pullrequestreview-1910510063, I believe it is unnecessary to bring in non-RDF references. It is not meaningful for our readers, and it looks like an arbitrary choice from the _huge_ literature on graph theory in general.

(I understand that some of these references may have been relevant for some of the cited works, i.e., they are, sort of, "second level" references. But the only goal of this section is to show that the works relies on a bunch of works that have been done in the RDF community.)